Spatially Structured Variance Ratio (SSVR)

Calculates the Spatially Structured Variance Ratio (SSVR) from a variogram model (the result of gstat::vgm, `gstat::fit.variogram, or a data.frame like gstat vgm). The SSVR quantifies the proportion of variance that is spatially structured: SSVR = 1 - (nugget / sill), where "sill" is total sill (sum of psill components). The output is a single numeric value between 0 and 1, where values closer to 1 indicate a higher proportion of the data explained by the spatial component.

vgm_ssvr(model)

Arguments

model

A variogram model as a data.frame with columns 'model' and 'psill', typically obtained from gstat::vgm or gstat::fit.variogram.

Value

A numeric value between 0 and 1 representing the SSVR.

References

Kerry, R., & Oliver, M. A. (2008). Determining nugget: sill ratios of standardized variograms from aerial photographs to krige sparse soil data. Precision Agriculture, 9(1), 33-56.

Vaysse, K., & Lagacherie, P. (2015). Evaluating digital soil mapping approaches for mapping GlobalSoilMap soil properties from legacy data in Languedoc-Roussillon (France). Geoderma Regional, 4, 20-30.

Examples

vgm_model = data.frame(model = c("Nug", "Sph"), psill = c(0.5, 1.5), range = c(0, 100))
ssvr_value = vgm_ssvr(vgm_model)
ssvr_value
#> [1] 0.75